fix: ssr and csr client components
